docs: Add GitHub token setup instructions to part_01_intro README
This commit adds a new section to the part_01_intro/README.md
explaining how to set up a GitHub Personal Access Token to avoid
rate limiting when using the 'gh-release' feature of the z3 crate.
It also includes a minor formatting fix in Cargo.toml by adding a
newline at the end of the file, which was an unintended side effect
of a previous operation but is a good practice.